zuperscore.requirements.production.txt
asgiref==3.7.2
async-timeout==4.0.2
attrs==23.1.0
autopep8==2.0.2
boto3==1.28.20
botocore==1.31.20
cached-property==1.5.2
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==41.0.3
defusedxml==0.7.1
Deprecated==1.2.14
dj-database-url==0.5.0
Django==3.2.20
django-allauth==0.50.0
django-braces==1.15.0
django-cors-headers==3.11.0
django-crum==0.7.9
django-debug-toolbar==3.2.4
django-filter==21.1
django-nested-admin==3.4.0
django-oauth-toolkit==2.0.0
django-redis==5.2.0
django-rest-swagger==2.2.0
django-rq==2.7.0
django-s3-storage==0.13.6
django-silk==5.0.3
django-taggit==2.1.0
django-treebeard==4.5.1
django-url-filter==0.3.15
djangorestframework==3.13.1
djangorestframework-simplejwt==5.1.0
drf-yasg==1.21.7
enum-compat==0.0.3
Faker==13.4.0
gevent==22.10.1
gprof2dot==2022.7.29
greenlet==1.1.3.post0
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jmespath==1.0.1
jsonmodels==2.5.0
jsonschema==4.19.0
jsonschema-specifications==2023.7.1
jwcrypto==1.5.0
MarkupSafe==2.1.3
oauthlib==3.2.2
openapi-codec==1.3.2
packaging==23.1
Pillow==9.5.0
postgres==4.0
psycopg2==2.9.3
psycopg2-binary==2.9.7
psycopg2-pool==1.1
pycodestyle==2.11.0
pycparser==2.21
PyJWT==2.8.0
python-dateutil==2.8.2
python-dotenv==1.0.0
python-monkey-business==1.0.0
python3-openid==3.2.0
pytz==2023.3
PyYAML==6.0.1
redis==4.3.4
referencing==0.30.2
requests==2.31.0
requests-oauthlib==1.3.1
rpds-py==0.10.0
rq==1.15.1
s3transfer==0.6.1
sentry-sdk==1.5.12
simplejson==3.19.2
six==1.16.0
sqlparse==0.4.4
twilio==7.16.2
uritemplate==4.1.1
urllib3==1.26.16
whitenoise==6.0.0
wrapt==1.15.0
zope.event==5.0
zope.interface==6.0
